refactor(oauth): separate persistence stores from refresh policy
