+++ docs/refactor/dedupe-2026-08-01.md
@@ -14,7 +14,10 @@ Completed. The shared foundations now own stable JSON serialization, daemon conf
 
 Completed. Record and replay now share config writing, subprocess propagation, and environment instructions. Call, list, and auth render one structured ad-hoc help definition with their original alignment. The two intentionally distinct definition serializers share only their common field projection, and Chrome's standalone fallback is rendered from the canonical patch implementation.
 
-## Remaining commits
+## Commit 3: OAuth persistence and refresh policy
 
-1. Separate OAuth persistence stores from refresh policy.
-2. Isolate runtime connection cache and transport ownership, then delete this tracker.
+Completed. The public OAuth persistence facade now delegates to internal directory, vault, and composite stores, while expiry classification and both OAuth and refreshable-bearer refresh policy live in a dedicated module. Store snapshots reconcile identity once for read-many migration, and client-secret lookup is canonical without changing the stricter refresh-path handling of blank environment values.
+
+## Remaining commit
+
+Isolate runtime connection cache and transport ownership, then delete this tracker.
