+++ src/oauth-persistence.ts
@@ -1,43 +1,22 @@
 import fs from 'node:fs/promises';
-import { Buffer } from 'node:buffer';
-import os from 'node:os';
-import path from 'node:path';
-import { discoverOAuthServerInfo, refreshAuthorization } from '@modelcontextprotocol/sdk/client/auth.js';
-import type {
-  OAuthClientInformationMixed,
-  OAuthProtectedResourceMetadata,
-  OAuthTokens,
-} from '@modelcontextprotocol/sdk/shared/auth.js';
-import { checkResourceAllowed, resourceUrlFromServerUrl } from '@modelcontextprotocol/sdk/shared/auth-utils.js';
+import type { OAuthClientInformationMixed, OAuthTokens } from '@modelcontextprotocol/sdk/shared/auth.js';
 import type { ServerDefinition } from './config.js';
-import { readJsonFile, withFileLock, writeJsonFile, writeTextFileAtomic } from './fs-json.js';
 import type { Logger } from './logging.js';
-import { buildStaticClientInformation } from './oauth-client-info.js';
-import {
-  sameOAuthClientGeneration,
-  sameOAuthClientValue,
-  sameOAuthTokenGeneration,
-  sameOAuthTokenValue,
-  withHiddenOAuthClientGeneration,
-  withHiddenOAuthTokenGeneration,
-  withOAuthClientGeneration,
-  withOAuthTokenGeneration,
-} from './oauth-token-generation.js';
-import {
-  clearVaultEntry,
-  clearVaultTokensIfMatching,
-  getOAuthVaultPath,
-  loadVaultEntry,
-  loadVaultEntryForRecovery,
-  reconcileVaultServerUrl,
-  saveVaultEntry,
-} from './oauth-vault.js';
-import { legacyMcporterDir } from './paths.js';
+import { clearLegacyOAuthArtifacts, createOAuthPersistenceStores } from './oauth-persistence-stores.js';
+import { readCachedAccessTokenWithPersistence } from './oauth-token-refresh.js';
 
 export type OAuthClearScope = 'all' | 'client' | 'tokens' | 'verifier' | 'state';
 
+export interface OAuthPersistenceSnapshot {
+  readonly tokens?: OAuthTokens;
+  readonly clientInfo?: OAuthClientInformationMixed;
+  readonly codeVerifier?: string;
+  readonly state?: string;
+}
+
 export interface OAuthPersistence {
   describe(): string;
+  readSnapshot(): Promise<OAuthPersistenceSnapshot>;
   readTokens(): Promise<OAuthTokens | undefined>;
   saveTokens(tokens: OAuthTokens): Promise<void>;
   readClientInfo(): Promise<OAuthClientInformationMixed | undefined>;
@@ -55,563 +34,8 @@ export interface OAuthPersistence {
   ): Promise<void>;
 }
 
-type StoredOAuthTokens = OAuthTokens & {
-  expires_at?: number;
-  expiresAt?: number;
-};
-
-const TOKEN_EXPIRY_SKEW_SECONDS = 60;
-
-function withStoredExpiry(tokens: OAuthTokens): OAuthTokens {
-  const stored = tokens as StoredOAuthTokens;
-  if (typeof stored.expires_at === 'number' || typeof stored.expiresAt === 'number') {
-    return tokens;
-  }
-  if (typeof tokens.expires_in === 'number' && Number.isFinite(tokens.expires_in)) {
-    return {
-      ...tokens,
-      expires_at: Math.floor(Date.now() / 1000) + tokens.expires_in,
-    } as OAuthTokens;
-  }
-  return tokens;
-}
-
-function prepareStoredTokens(tokens: OAuthTokens): OAuthTokens {
-  return withStoredExpiry(withOAuthTokenGeneration(tokens));
-}
-
-function tokenExpirySeconds(tokens: OAuthTokens): number | undefined {
-  const stored = tokens as StoredOAuthTokens;
-  for (const candidate of [stored.expires_at, stored.expiresAt]) {
-    if (typeof candidate === 'number' && Number.isFinite(candidate)) {
-      return candidate;
-    }
-  }
-  return undefined;
-}
-
-function cachedTokensChanged(original: OAuthTokens, current: OAuthTokens | undefined): boolean {
-  if (!current || typeof current.access_token !== 'string' || current.access_token.trim().length === 0) {
-    return false;
-  }
-  return !sameOAuthTokenGeneration(current, original);
-}
-
-function shouldRefreshCachedToken(tokens: OAuthTokens, skewSeconds = TOKEN_EXPIRY_SKEW_SECONDS): boolean {
-  const expiresAt = tokenExpirySeconds(tokens);
-  if (expiresAt !== undefined) {
-    return expiresAt <= Math.floor(Date.now() / 1000) + skewSeconds;
-  }
-  return typeof tokens.expires_in === 'number' && typeof tokens.refresh_token === 'string';
-}
-
-function resourceForRefresh(
-  serverUrl: URL,
-  resourceMetadata: OAuthProtectedResourceMetadata | undefined
-): URL | undefined {
-  if (!resourceMetadata) {
-    return undefined;
-  }
-  const defaultResource = resourceUrlFromServerUrl(serverUrl);
-  if (!checkResourceAllowed({ requestedResource: defaultResource, configuredResource: resourceMetadata.resource })) {
-    throw new Error(
-      `Protected resource ${resourceMetadata.resource} does not match expected ${defaultResource} (or origin)`
-    );
-  }
-  return new URL(resourceMetadata.resource);
-}
-
-function unrecoverableOAuthRefreshCode(error: unknown): string | undefined {
-  const errorCode = oauthErrorCode(error);
-  if (errorCode && ['invalid_client', 'invalid_grant', 'unauthorized_client'].includes(errorCode)) {
-    return errorCode;
-  }
-  return undefined;
-}
-
-async function oauthErrorCodeFromResponse(response: Response): Promise<string | undefined> {
-  try {
-    const payload = (await response.json()) as { error?: unknown };
-    if (payload && typeof payload.error === 'string' && payload.error.length > 0) {
-      return payload.error.toLowerCase();
-    }
-  } catch {
-    // Non-JSON error body; leave the error unclassified.
-  }
-  return undefined;
-}
-
-function oauthErrorCode(error: unknown): string | undefined {
-  if (!error || typeof error !== 'object') {
-    return undefined;
-  }
-  const { errorCode, name } = error as { errorCode?: unknown; name?: unknown };
-  if (typeof errorCode === 'string' && errorCode.length > 0) {
-    return errorCode.toLowerCase();
-  }
-  if (typeof name === 'string') {
-    const normalized = name.toLowerCase();
-    if (normalized === 'invalidclienterror') {
-      return 'invalid_client';
-    }
-    if (normalized === 'invalidgranterror') {
-      return 'invalid_grant';
-    }
-    if (normalized === 'unauthorizedclienterror') {
-      return 'unauthorized_client';
-    }
-  }
-  return undefined;
-}
-
-class DirectoryPersistence implements OAuthPersistence {
-  private readonly tokenPath: string;
-  private readonly clientInfoPath: string;
-  private readonly codeVerifierPath: string;
-  private readonly statePath: string;
-  private readonly serverUrlPath: string;
-
-  constructor(
-    private readonly root: string,
-    private readonly logger?: Logger,
-    private readonly serverUrl?: string,
-    private readonly skipUrlMarkerWhenMissing = false
-  ) {
-    this.tokenPath = path.join(root, 'tokens.json');
-    this.clientInfoPath = path.join(root, 'client.json');
-    this.codeVerifierPath = path.join(root, 'code_verifier.txt');
-    this.statePath = path.join(root, 'state.txt');
-    this.serverUrlPath = path.join(root, 'server_url.txt');
-  }
-
-  describe(): string {
-    return this.root;
-  }
-
-  private async ensureDir() {
-    await fs.mkdir(this.root, { recursive: true });
-  }
-
-  private async reconcileServerUrl(): Promise<void> {
-    if (!this.serverUrl) {
-      return;
-    }
-    const serverUrl = this.serverUrl;
-    if (this.skipUrlMarkerWhenMissing) {
-      try {
-        await fs.access(this.root);
-      } catch (error) {
-        if ((error as NodeJS.ErrnoException).code === 'ENOENT') {
-          return;
-        }
-        throw error;
-      }
-    }
-    await this.ensureDir();
-    await withFileLock(this.tokenPath, async () => {
-      let previousUrl: string | undefined;
-      try {
-        previousUrl = (await fs.readFile(this.serverUrlPath, 'utf8')).trim();
-      } catch (error) {
-        if ((error as NodeJS.ErrnoException).code !== 'ENOENT') {
-          throw error;
-        }
-      }
-      if (previousUrl === serverUrl) {
-        return;
-      }
-      // Keep the marker after invalidation so A -> B -> A cannot revive A's
-      // old directory-backed credentials.
-      if (previousUrl !== undefined) {
-        await this.clearFiles('all');
-      }
-      await writeTextFileAtomic(this.serverUrlPath, serverUrl);
-    });
-  }
-
-  async readTokens(): Promise<OAuthTokens | undefined> {
-    await this.reconcileServerUrl();
-    const tokens = await this.readJsonOrUndefined<OAuthTokens>(this.tokenPath);
-    return tokens ? withHiddenOAuthTokenGeneration(tokens) : undefined;
-  }
-
-  async saveTokens(tokens: OAuthTokens): Promise<void> {
-    await this.reconcileServerUrl();
-    await this.ensureDir();
-    // Locked so clearRejectedCredentials cannot compare-then-unlink across a
-    // concurrent write.
-    await withFileLock(this.tokenPath, async () => {
-      await writeJsonFile(this.tokenPath, prepareStoredTokens(tokens));
-    });
-    this.logger?.debug?.(`Saved tokens to ${this.tokenPath}`);
-  }
-
-  async clearRejectedCredentials(
-    expectedTokens?: OAuthTokens,
-    expectedClientInfo?: OAuthClientInformationMixed
-  ): Promise<void> {
-    await this.reconcileServerUrl();
-    await withFileLock(this.tokenPath, async () => {
-      if (expectedTokens) {
-        const current = await this.readJsonOrUndefined<OAuthTokens>(this.tokenPath);
-        if (sameOAuthTokenGeneration(current, expectedTokens)) {
-          await this.unlinkIfPresent(this.tokenPath);
-        }
-      }
-      if (expectedClientInfo) {
-        const currentClientInfo = await this.readJsonOrUndefined<OAuthClientInformationMixed>(this.clientInfoPath);
-        if (sameOAuthClientGeneration(currentClientInfo, expectedClientInfo)) {
-          await this.unlinkIfPresent(this.clientInfoPath);
-        }
-      }
-    });
-  }
-
-  private async unlinkIfPresent(file: string): Promise<void> {
-    try {
-      await fs.unlink(file);
-    } catch (error) {
-      if ((error as NodeJS.ErrnoException).code !== 'ENOENT') {
-        throw error;
-      }
-    }
-  }
-
-  async readClientInfo(): Promise<OAuthClientInformationMixed | undefined> {
-    await this.reconcileServerUrl();
-    const info = await this.readJsonOrUndefined<OAuthClientInformationMixed>(this.clientInfoPath);
-    return info ? withHiddenOAuthClientGeneration(info) : undefined;
-  }
-
-  async saveClientInfo(info: OAuthClientInformationMixed): Promise<void> {
-    await this.reconcileServerUrl();
-    await this.ensureDir();
-    await withFileLock(this.tokenPath, async () => {
-      await writeJsonFile(this.clientInfoPath, withOAuthClientGeneration(info));
-    });
-  }
-
-  async readCodeVerifier(): Promise<string | undefined> {
-    await this.reconcileServerUrl();
-    try {
-      return (await fs.readFile(this.codeVerifierPath, 'utf8')).trim();
-    } catch (error) {
-      if ((error as NodeJS.ErrnoException).code === 'ENOENT') {
-        return undefined;
-      }
-      throw error;
-    }
-  }
-
-  async saveCodeVerifier(value: string): Promise<void> {
-    await this.reconcileServerUrl();
-    await this.ensureDir();
-    await writeTextFileAtomic(this.codeVerifierPath, value);
-  }
-
-  async readState(): Promise<string | undefined> {
-    await this.reconcileServerUrl();
-    // Deliberately NOT corrupt-tolerant: a corrupt OAuth state must fail the
-    // flow closed. Returning undefined here would skip the CSRF state check on
-    // the authorization callback (see oauth.ts), so only the credential caches
-    // (tokens/client) degrade to re-auth.
-    return readJsonFile<string>(this.statePath);
-  }
-
-  // A present-but-corrupt credential cache (tokens/client) means "no usable
-  // credentials": degrade to re-auth instead of crashing the connection,
-  // mirroring VaultPersistence and the daemon/server-proxy readers. Genuine I/O
-  // faults still propagate (readJsonFile re-throws everything except ENOENT).
-  // OAuth state is intentionally excluded (see readState) so its CSRF check
-  // still fails closed on a corrupt state file.
-  private async readJsonOrUndefined<T>(filePath: string): Promise<T | undefined> {
-    try {
-      return await readJsonFile<T>(filePath);
-    } catch (error) {
-      if (!(error instanceof SyntaxError)) {
-        throw error;
-      }
-      this.logger?.debug?.(`Ignoring corrupt OAuth cache file ${filePath}: ${error.message}`);
-      return undefined;
-    }
-  }
-
-  async saveState(value: string): Promise<void> {
-    await this.reconcileServerUrl();
-    await this.ensureDir();
-    await writeJsonFile(this.statePath, value);
-  }
-
-  async clear(scope: OAuthClearScope): Promise<void> {
-    await this.reconcileServerUrl();
-    await this.clearFiles(scope);
-  }
-
-  private async clearFiles(scope: OAuthClearScope): Promise<void> {
-    const files: string[] = [];
-    if (scope === 'all' || scope === 'tokens') {
-      files.push(this.tokenPath);
-    }
-    if (scope === 'all' || scope === 'client') {
-      files.push(this.clientInfoPath);
-    }
-    if (scope === 'all' || scope === 'verifier') {
-      files.push(this.codeVerifierPath);
-    }
-    if (scope === 'all' || scope === 'state') {
-      files.push(this.statePath);
-    }
-    await Promise.all(
-      files.map(async (file) => {
-        try {
-          await fs.unlink(file);
-        } catch (error) {
-          if ((error as NodeJS.ErrnoException).code !== 'ENOENT') {
-            throw error;
-          }
-        }
-      })
-    );
-  }
-}
-
-class VaultPersistence implements OAuthPersistence {
-  private tokenSnapshots: ReadonlyMap<string, OAuthTokens> | undefined;
-  private clientSnapshots: ReadonlyMap<string, OAuthClientInformationMixed> | undefined;
-
-  constructor(private readonly definition: ServerDefinition) {}
-
-  describe(): string {
-    return `${getOAuthVaultPath()} (vault)`;
-  }
-
-  private async reconcileServerUrl(): Promise<void> {
-    await reconcileVaultServerUrl(this.definition);
-  }
-
-  async readTokens(): Promise<OAuthTokens | undefined> {
-    await this.reconcileServerUrl();
-    const recovery = await loadVaultEntryForRecovery(this.definition);
-    this.tokenSnapshots = recovery.tokenSnapshots;
-    return recovery.entry?.tokens;
-  }
-
-  async saveTokens(tokens: OAuthTokens): Promise<void> {
-    await this.reconcileServerUrl();
-    await saveVaultEntry(this.definition, { tokens: prepareStoredTokens(tokens) });
-  }
-
-  async readClientInfo(): Promise<OAuthClientInformationMixed | undefined> {
-    await this.reconcileServerUrl();
-    const recovery = await loadVaultEntryForRecovery(this.definition);
-    this.clientSnapshots = recovery.clientSnapshots;
-    return recovery.entry?.clientInfo;
-  }
-
-  async saveClientInfo(info: OAuthClientInformationMixed): Promise<void> {
-    await this.reconcileServerUrl();
-    await saveVaultEntry(this.definition, { clientInfo: info });
-  }
-
-  async readCodeVerifier(): Promise<string | undefined> {
-    await this.reconcileServerUrl();
-    return (await loadVaultEntry(this.definition))?.codeVerifier;
-  }
-
-  async saveCodeVerifier(value: string): Promise<void> {
-    await this.reconcileServerUrl();
-    await saveVaultEntry(this.definition, { codeVerifier: value });
-  }
-
-  async readState(): Promise<string | undefined> {
-    await this.reconcileServerUrl();
-    return (await loadVaultEntry(this.definition))?.state;
-  }
-
-  async saveState(value: string): Promise<void> {
-    await this.reconcileServerUrl();
-    await saveVaultEntry(this.definition, { state: value });
-  }
-
-  async clear(scope: OAuthClearScope): Promise<void> {
-    await this.reconcileServerUrl();
-    await clearVaultEntry(this.definition, scope);
-  }
-
-  async clearRejectedCredentials(
-    expectedTokens?: OAuthTokens,
-    expectedClientInfo?: OAuthClientInformationMixed
-  ): Promise<void> {
-    await this.reconcileServerUrl();
-    await clearVaultTokensIfMatching(
-      this.definition,
-      expectedTokens,
-      expectedClientInfo,
-      expectedTokens ? this.tokenSnapshots : undefined,
-      expectedClientInfo ? this.clientSnapshots : undefined
-    );
-  }
-}
-
-class CompositePersistence implements OAuthPersistence {
-  private readonly recoveryTokenSnapshots = new Map<OAuthPersistence, OAuthTokens | undefined>();
-  private readonly recoveryClientSnapshots = new Map<OAuthPersistence, OAuthClientInformationMixed | undefined>();
-  private recoveryTokenSource: OAuthPersistence | undefined;
-  private recoveryClientSource: OAuthPersistence | undefined;
-
-  constructor(private readonly stores: OAuthPersistence[]) {}
-
-  describe(): string {
-    return this.stores.map((store) => store.describe()).join(' + ');
-  }
-
-  async readTokens(): Promise<OAuthTokens | undefined> {
-    const result = await this.readRecoveryValues((store) => store.readTokens(), this.recoveryTokenSnapshots);
-    this.recoveryTokenSource = result.source;
-    return result.value;
-  }
-
-  async saveTokens(tokens: OAuthTokens): Promise<void> {
-    // Compute the absolute expiry once so every backing store records the same
-    // generation value even at a wall-clock second boundary.
-    const stored = prepareStoredTokens(tokens);
-    await Promise.all(this.stores.map((store) => store.saveTokens(stored)));
-  }
-
-  async clearRejectedCredentials(
-    expectedTokens?: OAuthTokens,
-    expectedClientInfo?: OAuthClientInformationMixed
-  ): Promise<void> {
-    await Promise.all(
-      this.stores.map((store) => {
-        const storedTokenSnapshot = expectedTokens ? this.recoveryTokenSnapshots.get(store) : undefined;
-        const tokenSnapshot =
-          expectedTokens &&
-          storedTokenSnapshot &&
-          (store === this.recoveryTokenSource || sameOAuthTokenValue(storedTokenSnapshot, expectedTokens))
-            ? storedTokenSnapshot
-            : undefined;
-        const storedClientSnapshot = expectedClientInfo ? this.recoveryClientSnapshots.get(store) : undefined;
-        const clientSnapshot =
-          expectedClientInfo &&
-          storedClientSnapshot &&
-          (store === this.recoveryClientSource || sameOAuthClientValue(storedClientSnapshot, expectedClientInfo))
-            ? storedClientSnapshot
-            : undefined;
-        return tokenSnapshot || clientSnapshot
-          ? store.clearRejectedCredentials(tokenSnapshot, clientSnapshot)
-          : Promise.resolve();
-      })
-    );
-  }
-
-  async readClientInfo(): Promise<OAuthClientInformationMixed | undefined> {
-    const result = await this.readRecoveryValues((store) => store.readClientInfo(), this.recoveryClientSnapshots);
-    this.recoveryClientSource = result.source;
-    return result.value;
-  }
-
-  private async readRecoveryValues<T>(
-    read: (store: OAuthPersistence) => Promise<T | undefined>,
-    snapshots: Map<OAuthPersistence, T | undefined>
-  ): Promise<{ value: T | undefined; source: OAuthPersistence | undefined }> {
-    const results = await Promise.allSettled(this.stores.map((store) => read(store)));
-    snapshots.clear();
-    let value: T | undefined;
-    let source: OAuthPersistence | undefined;
-    for (const [index, result] of results.entries()) {
-      const store = this.stores[index]!;
-      if (result.status === 'rejected') {
-        // Preserve ordered fallback semantics: a lower-priority store cannot
-        // invalidate an already-readable primary cache, while failures before
-        // the first usable value still surface.
-        if (!source) {
-          throw result.reason;
-        }
-        continue;
-      }
-      snapshots.set(store, result.value);
-      if (!source && result.value !== undefined) {
-        source = store;
-        value = result.value;
-      }
-    }
-    return { value, source };
-  }
-
-  async saveClientInfo(info: OAuthClientInformationMixed): Promise<void> {
-    const stored = withOAuthClientGeneration(info);
-    await Promise.all(this.stores.map((store) => store.saveClientInfo(stored)));
-  }
-
-  async readCodeVerifier(): Promise<string | undefined> {
-    for (const store of this.stores) {
-      const result = await store.readCodeVerifier();
-      if (result) {
-        return result;
-      }
-    }
-    return undefined;
-  }
-
-  async saveCodeVerifier(value: string): Promise<void> {
-    await Promise.all(this.stores.map((store) => store.saveCodeVerifier(value)));
-  }
-
-  async readState(): Promise<string | undefined> {
-    for (const store of this.stores) {
-      const result = await store.readState();
-      if (result) {
-        return result;
-      }
-    }
-    return undefined;
-  }
-
-  async saveState(value: string): Promise<void> {
-    await Promise.all(this.stores.map((store) => store.saveState(value)));
-  }
-
-  async clear(scope: OAuthClearScope): Promise<void> {
-    await Promise.all(this.stores.map((store) => store.clear(scope)));
-  }
-}
-
 export async function buildOAuthPersistence(definition: ServerDefinition, logger?: Logger): Promise<OAuthPersistence> {
-  const vault = new VaultPersistence(definition);
-  const stores: OAuthPersistence[] = [vault];
-  const serverUrl = definition.command.kind === 'http' ? definition.command.url.toString() : undefined;
-
-  if (definition.tokenCacheDir) {
-    stores.unshift(new DirectoryPersistence(definition.tokenCacheDir, logger, serverUrl));
-  }
-
-  // Migrate legacy default per-server cache (~/.mcporter/<name>) into the vault if present.
-  const legacyDir = path.join(legacyMcporterDir(), definition.name);
-  if (!definition.tokenCacheDir) {
-    const legacy = new DirectoryPersistence(legacyDir, logger, serverUrl, true);
-    const legacyTokens = await legacy.readTokens();
-    const legacyClient = await legacy.readClientInfo();
-    const legacyVerifier = await legacy.readCodeVerifier();
-    const legacyState = await legacy.readState();
-    if (legacyTokens || legacyClient || legacyVerifier || legacyState) {
-      if (legacyTokens) {
-        await vault.saveTokens(legacyTokens);
-      }
-      if (legacyClient) {
-        await vault.saveClientInfo(legacyClient);
-      }
-      if (legacyVerifier) {
-        await vault.saveCodeVerifier(legacyVerifier);
-      }
-      if (legacyState) {
-        await vault.saveState(legacyState);
-      }
-      logger?.info?.(`Migrated legacy OAuth cache for '${definition.name}' into vault.`);
-    }
-  }
-
-  return stores.length === 1 ? vault : new CompositePersistence(stores);
+  return await createOAuthPersistenceStores(definition, logger);
 }
 
 export async function clearOAuthCaches(
@@ -629,318 +53,10 @@ export async function clearOAuthCaches(
   await clearLegacyOAuthArtifacts(definition, logger, scope);
 }
 
-// Legacy artifacts are never written by live refresh winners (the migration in
-// buildOAuthPersistence only reads them), so clearing them cannot race a
-// concurrent token save.
-async function clearLegacyOAuthArtifacts(
-  definition: ServerDefinition,
-  logger: Logger | undefined,
-  scope: OAuthClearScope
-): Promise<void> {
-  const legacyDir = path.join(legacyMcporterDir(), definition.name);
-  if (!definition.tokenCacheDir || legacyDir !== definition.tokenCacheDir) {
-    const legacy = new DirectoryPersistence(legacyDir, logger);
-    await legacy.clear(scope);
-  }
-
-  // Known provider-specific legacy paths (gmail server writes to ~/.gmail-mcp/credentials.json).
-  const legacyFiles: string[] = [];
-  if (definition.name.toLowerCase() === 'gmail') {
-    legacyFiles.push(path.join(os.homedir(), '.gmail-mcp', 'credentials.json'));
-  }
-  await Promise.all(
-    legacyFiles.map(async (file) => {
-      try {
-        await fs.unlink(file);
-        logger?.info?.(`Cleared legacy OAuth cache file ${file}`);
-      } catch (error) {
-        if ((error as NodeJS.ErrnoException).code !== 'ENOENT') {
-          throw error;
-        }
-      }
-    })
-  );
-}
-
 export async function readCachedAccessToken(
   definition: ServerDefinition,
   logger?: Logger
 ): Promise<string | undefined> {
   const persistence = await buildOAuthPersistence(definition, logger);
-  const tokens = await persistence.readTokens();
-  if (!tokens || typeof tokens.access_token !== 'string' || tokens.access_token.trim().length === 0) {
-    return undefined;
-  }
-  if (definition.auth === 'refreshable_bearer') {
-    return await readExplicitRefreshableBearerToken(definition, persistence, tokens, logger);
-  }
-  if (!shouldRefreshCachedToken(tokens)) {
-    return tokens.access_token;
-  }
-  if (typeof tokens.refresh_token !== 'string' || tokens.refresh_token.trim().length === 0) {
-    return tokens.access_token;
-  }
-  let persistedClientInformation: OAuthClientInformationMixed | undefined;
-  try {
-    const staticClientInformation = buildStaticClientInformation(definition);
-    persistedClientInformation = staticClientInformation ? undefined : await persistence.readClientInfo();
-    const clientInformation = staticClientInformation ?? persistedClientInformation;
-    if (!clientInformation) {
-      logger?.debug?.(
-        `Cached OAuth token for '${definition.name}' is expired, but no client information is available.`
-      );
-      return tokens.access_token;
-    }
-    if (definition.command.kind !== 'http') {
-      return tokens.access_token;
-    }
-    const serverInfo = await discoverOAuthServerInfo(definition.command.url);
-    const resource = resourceForRefresh(definition.command.url, serverInfo.resourceMetadata);
-    const refreshed = await refreshAuthorization(serverInfo.authorizationServerUrl, {
-      metadata: serverInfo.authorizationServerMetadata,
-      clientInformation,
-      refreshToken: tokens.refresh_token,
-      ...(resource ? { resource } : {}),
-    });
-    await persistence.saveTokens(refreshed);
-    logger?.debug?.(`Refreshed cached OAuth access token for '${definition.name}' (non-interactive).`);
-    return refreshed.access_token;
-  } catch (error) {
-    logger?.debug?.(
-      `Failed to refresh cached OAuth token for '${definition.name}' non-interactively: ${
-        error instanceof Error ? error.message : String(error)
-      }`
-    );
-    const unrecoverableCode = unrecoverableOAuthRefreshCode(error);
-    if (unrecoverableCode) {
-      return await recoverFromUnrecoverableRefresh(
-        definition,
-        persistence,
-        tokens,
-        unrecoverableCode,
-        'OAuth',
-        logger,
-        persistedClientInformation
-      );
-    }
-    return tokens.access_token;
-  }
-}
-
-/**
- * After a refresh fails with an unrecoverable OAuth error code, adopt the
- * tokens a concurrent refresh persisted first, or clear the dead credentials
- * so the rejected refresh token is never replayed (providers that rotate
- * refresh tokens treat replays as stolen-token signals and revoke the grant).
- * Returns the concurrent winner's access token, or undefined after clearing.
- *
- * The rejected token and the specific dynamic client registration used by the
- * refresh are compare-and-cleared under each store's lock. State and PKCE
- * verifier data belong to interactive auth and are never removed here. There
- * is no later unconditional live-store sweep, so any different generation is
- * preserved and adopted.
- */
-async function recoverFromUnrecoverableRefresh(
-  definition: ServerDefinition,
-  persistence: OAuthPersistence,
-  tokens: OAuthTokens,
-  unrecoverableCode: string,
-  tokenKind: 'OAuth' | 'bearer',
-  logger?: Logger,
-  rejectedClientInformation?: OAuthClientInformationMixed
-): Promise<string | undefined> {
-  const clientToClear = unrecoverableCode === 'invalid_grant' ? undefined : rejectedClientInformation;
-  await persistence.clearRejectedCredentials(tokens, clientToClear);
-  // Legacy artifacts are not live refresh destinations, so this sweep cannot
-  // race a winner. Live stores were handled only by guarded mutations above.
-  await clearLegacyOAuthArtifacts(definition, logger, unrecoverableCode === 'invalid_grant' ? 'tokens' : 'all');
-  const latestTokens = await persistence.readTokens();
-  if (latestTokens && cachedTokensChanged(tokens, latestTokens)) {
-    logger?.debug?.(
-      `Kept cached ${tokenKind} token for '${definition.name}' because another refresh updated it first.`
-    );
-    return latestTokens.access_token;
-  }
-  logger?.debug?.(
-    `Cleared cached ${tokenKind} ${unrecoverableCode === 'invalid_grant' ? 'token' : 'credentials'} for '${
-      definition.name
-    }' after unrecoverable refresh failure.`
-  );
-  return undefined;
-}
-
-async function readExplicitRefreshableBearerToken(
-  definition: ServerDefinition,
-  persistence: OAuthPersistence,
-  tokens: OAuthTokens,
-  logger?: Logger
-): Promise<string> {
-  const refresh = definition.refresh;
-  const skewSeconds = refresh?.refreshSkewSeconds ?? TOKEN_EXPIRY_SKEW_SECONDS;
-  if (!shouldRefreshCachedToken(tokens, skewSeconds)) {
-    return tokens.access_token;
-  }
-  if (!refresh) {
-    throw new Error(`Cached bearer token for '${definition.name}' is expired, but refresh is not configured.`);
-  }
-  if (typeof tokens.refresh_token !== 'string' || tokens.refresh_token.trim().length === 0) {
-    throw new Error(`Cached bearer token for '${definition.name}' is expired, but no refresh_token is available.`);
-  }
-  try {
-    const refreshed = await refreshBearerToken(definition, tokens.refresh_token);
-    await persistence.saveTokens(refreshed);
-    logger?.debug?.(`Refreshed bearer access token for '${definition.name}' (non-interactive).`);
-    return refreshed.access_token;
-  } catch (error) {
-    logger?.debug?.(
-      `Failed to refresh bearer token for '${definition.name}' non-interactively: ${
-        error instanceof Error ? error.message : String(error)
-      }`
-    );
-    const unrecoverableCode = unrecoverableOAuthRefreshCode(error);
-    if (unrecoverableCode) {
-      const winnerToken = await recoverFromUnrecoverableRefresh(
-        definition,
-        persistence,
-        tokens,
-        unrecoverableCode,
-        'bearer',
-        logger
-      );
-      if (winnerToken) {
-        return winnerToken;
-      }
-    }
-    throw new Error(
-      `Failed to refresh cached bearer token for '${definition.name}': ${
-        error instanceof Error ? error.message : String(error)
-      }`,
-      { cause: error }
-    );
-  }
-}
-
-async function refreshBearerToken(definition: ServerDefinition, refreshToken: string): Promise<OAuthTokens> {
-  const refresh = definition.refresh;
-  if (!refresh) {
-    throw new Error('Missing refresh configuration.');
-  }
-  const clientId = readEnvOrConfig(refresh.clientIdEnv, definition.oauthClientId);
-  const method = refresh.clientAuthMethod ?? definition.oauthTokenEndpointAuthMethod ?? 'client_secret_basic';
-  const clientSecret = method === 'none' ? undefined : readClientSecret(definition, refresh.clientSecretEnv);
-  const body = new URLSearchParams({
-    grant_type: 'refresh_token',
-    refresh_token: refreshToken,
-  });
-  const headers: Record<string, string> = {
-    accept: 'application/json',
-    'content-type': 'application/x-www-form-urlencoded',
-  };
-
-  if (method === 'client_secret_post') {
-    if (clientId) {
-      body.set('client_id', clientId);
-    }
-    if (clientSecret) {
-      body.set('client_secret', clientSecret);
-    }
-  } else if (method === 'none') {
-    if (clientId) {
-      body.set('client_id', clientId);
-    }
-  } else {
-    if (!clientId || !clientSecret) {
-      throw new Error(`Refresh client credentials are required for '${method}'.`);
-    }
-    headers.authorization = `Basic ${Buffer.from(
-      `${formEncodeCredential(clientId)}:${formEncodeCredential(clientSecret)}`
-    ).toString('base64')}`;
-  }
-
-  const response = await fetch(refresh.tokenEndpoint, {
-    method: 'POST',
-    headers,
-    body,
-  });
-  if (!response.ok) {
-    const errorCode = await oauthErrorCodeFromResponse(response);
-    throw Object.assign(
-      new Error(`Token endpoint returned HTTP ${response.status}${errorCode ? ` (${errorCode})` : ''}.`),
-      errorCode ? { errorCode } : {}
-    );
-  }
-  const payload = normalizeBearerTokenResponse(await response.json());
-  return {
-    ...payload,
-    ...(payload.refresh_token ? {} : { refresh_token: refreshToken }),
-  };
-}
-
-function normalizeBearerTokenResponse(value: unknown): OAuthTokens {
-  if (!value || typeof value !== 'object') {
-    throw new Error('Token endpoint did not return a JSON object.');
-  }
-  const payload = value as Record<string, unknown>;
-  if (typeof payload.access_token !== 'string' || payload.access_token.trim().length === 0) {
-    throw new Error('Token endpoint did not return an access_token.');
-  }
-  return {
-    access_token: payload.access_token,
-    token_type: typeof payload.token_type === 'string' && payload.token_type ? payload.token_type : 'Bearer',
-    ...(typeof payload.id_token === 'string' ? { id_token: payload.id_token } : {}),
-    ...(typeof payload.scope === 'string' ? { scope: payload.scope } : {}),
-    ...(typeof payload.refresh_token === 'string' && payload.refresh_token
-      ? { refresh_token: payload.refresh_token }
-      : {}),
-    ...coerceExpiresIn(payload.expires_in),
-  };
-}
-
-function coerceExpiresIn(value: unknown): Pick<OAuthTokens, 'expires_in'> {
-  if (typeof value === 'number' && Number.isFinite(value)) {
-    return { expires_in: value };
-  }
-  if (typeof value === 'string' && value.trim().length > 0) {
-    const parsed = Number(value);
-    if (Number.isFinite(parsed)) {
-      return { expires_in: parsed };
-    }
-  }
-  return {};
-}
-
-function readEnvOrConfig(envName: string | undefined, fallback: string | undefined): string | undefined {
-  if (!envName) {
-    return fallback;
-  }
-  const value = process.env[envName];
-  if (value === undefined || value.trim().length === 0) {
-    throw new Error(`Environment variable '${envName}' is required for bearer token refresh.`);
-  }
-  return value;
-}
-
-function formEncodeCredential(value: string): string {
-  return new URLSearchParams([['', value]]).toString().slice(1);
-}
-
-function readClientSecret(
-  definition: ServerDefinition,
-  refreshClientSecretEnv: string | undefined
-): string | undefined {
-  if (refreshClientSecretEnv) {
-    return readEnvOrConfig(refreshClientSecretEnv, undefined);
-  }
-  return resolveOAuthClientSecret(definition);
-}
-
-function resolveOAuthClientSecret(definition: ServerDefinition): string | undefined {
-  if (definition.oauthClientSecretEnv) {
-    const value = process.env[definition.oauthClientSecretEnv];
-    if (value === undefined || value.trim().length === 0) {
-      throw new Error(`Environment variable '${definition.oauthClientSecretEnv}' is required for OAuth client secret.`);
-    }
-    return value;
-  }
-  return definition.oauthClientSecret;
+  return await readCachedAccessTokenWithPersistence(definition, persistence, logger);
 }
